+++ blocks/reader-revenue-manager/block-editor-plugin/editor-styles.scss
@@ -1,0 +1,29 @@
+/**
+ * Reader Revenue Manager: Editor Styles
+ *
+ * Site Kit by Google, Copyright 2025 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+.editor-sidebar {
+	.googlesitekit-rrm-settings-panel {
+
+		.googlesitekit-rrm-panel__select-control {
+			label {
+				text-transform: none;
+				white-space: break-spaces;
+			}
+		}
+	}
+}
