+++ blocks/reader-revenue-manager/block-editor-plugin/plugin-registration.js
@@ -26,14 +26,11 @@ import { registerPlugin } from '@wordpress-core/plugins';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import { registerStore } from '../../assets/js/modules/reader-revenue-manager/datastore';
-import { CORE_MODULES } from '../../assets/js/googlesitekit/modules/datastore/constants';
-import SettingPanel from './components/SettingPanel';
-import { MODULES_READER_REVENUE_MANAGER } from '../../assets/js/modules/reader-revenue-manager/datastore/constants';
+import { CORE_MODULES } from '../../../assets/js/googlesitekit/modules/datastore/constants';
+import SettingPanel from './SettingPanel';
+import { MODULES_READER_REVENUE_MANAGER } from '../../../assets/js/modules/reader-revenue-manager/datastore/constants';
 
 const { select, resolveSelect } = Data;
-
-registerStore( Data );
 
 export function registerReaderRevenueManagerPlugin() {
 	return Promise.all( [
